+++ tests/test_poetry_version_plugin.py
@@ -21,16 +21,21 @@ CONF = """
 [tool.poetry-version-plugin]
 source = "init"
 """
+CONF_2 = """
+
+[tool.poetry-dynamic-versioning]
+enable = true
+"""
 
 
 @contextmanager
 def _prepare_package(
-    package_path: Path, define_include=False
+    package_path: Path, define_include=False, mark="0"
 ) -> Generator[Path, None, None]:
     toml_file = package_path / TOML_FILE
     package_name = package_path.name.replace(" ", "_")
     init_file = package_path / package_name / "__init__.py"
-    a, b = 'version = "0.1.0"', 'version = "0"'
+    a, b = 'version = "0.1.0"', f'version = "{mark}"'
     if define_include:
         b += '\npackages = [{include = "%s"}]' % package_name
     with chdir(package_path.parent):
@@ -62,8 +67,27 @@ def test_version_plugin(tmp_path: Path) -> None:
             BumpUp(part="patch", commit=False, dry=True).gen()
 
 
+def test_version_plugin_2(tmp_path: Path) -> None:
+    with _prepare_package(tmp_path / "helloworld", mark="0.0.0") as init_file:
+        command = _build_bump_cmd(init_file)
+        assert BumpUp(part="patch", commit=False, dry=True).gen() == command
+        run_and_echo("poetry run fast bump patch")
+        assert init_file.read_text() == '__version__ = "0.0.2"\n'
+        init_file.unlink()
+        with pytest.raises(ParseError, match=r"Version file not found!.*"):
+            BumpUp(part="patch", commit=False, dry=True).gen()
+
+
 def test_version_plugin_include_defined(tmp_path: Path) -> None:
     with _prepare_package(tmp_path / "hello world", True) as init_file:
+        command = _build_bump_cmd(init_file)
+        assert BumpUp(part="patch", commit=False, dry=True).gen() == command
+        run_and_echo("poetry run fast bump patch")
+        assert init_file.read_text() == '__version__ = "0.0.2"\n'
+
+
+def test_version_plugin_include_defined_2(tmp_path: Path) -> None:
+    with _prepare_package(tmp_path / "hello world", True, mark="0.0.0") as init_file:
         command = _build_bump_cmd(init_file)
         assert BumpUp(part="patch", commit=False, dry=True).gen() == command
         run_and_echo("poetry run fast bump patch")